+++ crest_data_systems_netapp_ontap/datadog_checks/crest_data_systems_netapp_ontap/NetApp/NaServer.py
@@ -318,7 +318,7 @@ class NaServer:
             )
 
             if debug_style == "NA_PRINT_DONT_PARSE":
-                self.log.info("INPUT \n{content}".format(content=content))  # noqa: G00
+                self.log.info("INPUT \n{content}".format(content=content))  # noqa: G001
 
             connection.putheader("Content-length", str(len(content)))
             connection.endheaders()