+++ crest_data_systems_netapp_ontap/datadog_checks/crest_data_systems_netapp_ontap/cds_netapp_ontap_volume_perf.py
@@ -14,13 +14,13 @@ class VolumePerfIngestor:
         # collect performance data
         try:
             response = self.perf_handler_obj.collect(self.client)
-            avg_latency = list()
+            avg_latency = []
 
             if not response or not response.get("instances"):
                 self.log.info("NETAPP ONTAP INFO: Nothing to ingest in VOLUME performance data.")
                 return
 
-            instances = response["instances"].get("instance-data", list())
+            instances = response["instances"].get("instance-data", [])
             if isinstance(instances, dict):
                 instances = [instances]
 
@@ -28,7 +28,7 @@ class VolumePerfIngestor:
                 uuid = instance.get("uuid", "")
                 name = instance.get("name", "")
 
-                event = perf_response_parser(instance.get("counters", dict()).get("counter-data", list()))
+                event = perf_response_parser(instance.get("counters", {}).get("counter-data", []))
 
                 # ingest metrics of Block Operations
                 ingest_metric(