+++ crest_data_systems_netapp_ontap/CHANGELOG.md
@@ -1,4 +1,9 @@
 # CHANGELOG - crest_data_systems_netapp_ontap
 
+## 1.0.1 / 2023-04-18
+
+- [Updated] Support and setup information in readme.
+
 ## 1.0.0 / 2023-01-24
-- [Added] Initial release.
+
+- [Added] NetApp OnTap cluster version <=`NetApp Release 9.5P6` and NetApp OnTap 7-mode version <=`NetApp Release 8.2.4 7-Mode` are supported in this release.